+++ CHANGES.md
@@ -6,6 +6,8 @@
   - https://github.com/koyeb/koyeb-cli/pull/149
 * The new flag `--override` of `koyeb service update` allows to override the service configuration instead of merging the other options provided with the existing configuration.
   - https://github.com/koyeb/koyeb-cli/issues/147
+* Support `koyeb secret show <secret_name>`. *The API behind this command is not stable yet and may change in the future. This command can break at any time and will require an update of the CLI.*
+  - https://github.com/koyeb/koyeb-cli/issues/150
 
 **v3.3.2**